+++ wemake_python_styleguide/visitors/ast/functions.py
@@ -2,7 +2,7 @@
 
 import ast
 from contextlib import suppress
-from typing import ClassVar, Dict, List, Union
+from typing import ClassVar, Dict, List, Mapping, Union
 
 from typing_extensions import final
 
@@ -12,7 +12,7 @@ from wemake_python_styleguide.constants import (
     FUNCTIONS_BLACKLIST,
     LITERALS_BLACKLIST,
 )
-from wemake_python_styleguide.logic import nodes, walk
+from wemake_python_styleguide.logic import nodes, source, walk
 from wemake_python_styleguide.logic.arguments import function_args
 from wemake_python_styleguide.logic.naming import access
 from wemake_python_styleguide.logic.tree import (
@@ -47,6 +47,9 @@ from wemake_python_styleguide.visitors import base, decorators
 #: Things we treat as local variables.
 _LocalVariable = Union[ast.Name, ast.ExceptHandler]
 
+#: Function definitions with name and arity:
+_Defs = Mapping[str, int]
+
 
 @final
 class WrongFunctionCallVisitor(base.BaseNodeVisitor):
@@ -55,6 +58,21 @@ class WrongFunctionCallVisitor(base.BaseNodeVisitor):
 
     All these functions are defined in ``FUNCTIONS_BLACKLIST``.
     """
+
+    _functions: ClassVar[_Defs] = {
+        'getattr': 3,
+        'setattr': 3,
+    }
+
+    _postfixes: ClassVar[_Defs] = {
+        # dict methods:
+        '.get': 2,
+        '.pop': 2,
+        '.setdefault': 2,
+
+        # list methods:
+        '.insert': 2,
+    }
 
     def visit_Call(self, node: ast.Call) -> None:
         """
@@ -70,8 +88,12 @@ class WrongFunctionCallVisitor(base.BaseNodeVisitor):
         """
         self._check_wrong_function_called(node)
         self._check_boolean_arguments(node)
-        self._check_super_call(node)
         self._check_isinstance_call(node)
+
+        if functions.given_function_called(node, {'super'}):
+            self._check_super_context(node)
+            self._check_super_arguments(node)
+
         self.generic_visit(node)
 
     def _check_wrong_function_called(self, node: ast.Call) -> None:
@@ -91,15 +113,26 @@ class WrongFunctionCallVisitor(base.BaseNodeVisitor):
             if not isinstance(arg, ast.NameConstant):
                 continue
 
+            is_ignored = self._is_call_ignored(node)
+
             # We do not check for `None` values here:
-            if arg.value is True or arg.value is False:
+            if not is_ignored and arg.value in {True, False}:
                 self.add_violation(
                     BooleanPositionalArgumentViolation(
                         arg, text=str(arg.value),
                     ),
                 )
 
-    def _ensure_super_context(self, node: ast.Call) -> None:
+    def _check_isinstance_call(self, node: ast.Call) -> None:
+        function_name = functions.given_function_called(node, {'isinstance'})
+        if not function_name or len(node.args) != 2:
+            return
+
+        if isinstance(node.args[1], ast.Tuple):
+            if len(node.args[1].elts) == 1:
+                self.add_violation(WrongIsinstanceWithTupleViolation(node))
+
+    def _check_super_context(self, node: ast.Call) -> None:
         parent_context = nodes.get_context(node)
         parent_node = nodes.get_parent(node)
 
@@ -120,25 +153,24 @@ class WrongFunctionCallVisitor(base.BaseNodeVisitor):
             oop.WrongSuperCallViolation(node, text='not inside method'),
         )
 
-    def _ensure_super_arguments(self, node: ast.Call) -> None:
+    def _check_super_arguments(self, node: ast.Call) -> None:
         if node.args or node.keywords:
             self.add_violation(
                 oop.WrongSuperCallViolation(node, text='remove arguments'),
             )
 
-    def _check_super_call(self, node: ast.Call) -> None:
-        if functions.given_function_called(node, {'super'}):
-            self._ensure_super_context(node)
-            self._ensure_super_arguments(node)
+    def _is_call_ignored(self, node: ast.Call) -> bool:
+        call = source.node_to_string(node.func)
+        func_called = functions.given_function_called(
+            node, self._functions.keys(),
+        )
 
-    def _check_isinstance_call(self, node: ast.Call) -> None:
-        function_name = functions.given_function_called(node, {'isinstance'})
-        if not function_name or len(node.args) != 2:
-            return
-
-        if isinstance(node.args[1], ast.Tuple):
-            if len(node.args[1].elts) == 1:
-                self.add_violation(WrongIsinstanceWithTupleViolation(node))
+        return bool(
+            func_called and len(node.args) == self._functions[func_called],
+        ) or any(
+            call.endswith(post) and len(node.args) == self._postfixes[post]
+            for post in self._postfixes
+        )
 
 
 @final
@@ -297,7 +329,7 @@ class FunctionDefinitionVisitor(base.BaseNodeVisitor):
 
     def _ensure_used_variables(
         self,
-        local_variables: Dict[str, List[_LocalVariable]],
+        local_variables: Mapping[str, List[_LocalVariable]],
     ) -> None:
         for varname, usages in local_variables.items():
             for node in usages: